utils: sync, remove dead code
