+++ utils/evaluate_ref_models.m
@@ -5,9 +5,10 @@ addpath(fullfile(fileparts(mfilename('fullpath')), '..','examples', 'imagenet'))
 
 models = {...
   'matconvnet-alex', ...
-  'matconvnet-vgg-f', ...
-  'matconvnet-vgg-m', ...
   'matconvnet-vgg-s', ...
+  'matconvnet-vgg-m', ...
+  'matconvnet-vgg-f', ...
+  'matconvnet-vgg-verydeep-16', ...
   'caffe-ref', ...
   'caffe-alex', ...
   'vgg-s', ...
@@ -38,8 +39,8 @@ for i = 1:numel(models)
   end
 end
 
-fprintf('|%20s|%10s|%10s|%10s|\n', 'model', 'top-1 err.', 'top-5 err.', 'images/s') ;
-fprintf('%s\n', repmat('-',1,20+10+10+10+5)) ;
+fprintf('|%30s|%10s|%10s|%10s|\n', 'model', 'top-1 err.', 'top-5 err.', 'images/s') ;
+fprintf('%s\n', repmat('-',1,30+10+10+10+5)) ;
 
 for i = 1:numel(models)
   opts.expDir = fullfile('data', 'models-eval', models{i}) ;
@@ -56,7 +57,7 @@ for i = 1:numel(models)
     speed = results.val.num / results.val.time ;
   end
 
-  fprintf('|%20s|%10s|%10s|%10s|\n', ...
+  fprintf('|%30s|%10s|%10s|%10s|\n', ...
     models{i}, ...
     sprintf('%5.1f',top1*100), ...
     sprintf('%5.1f',top5*100), ...
