+++ utils/import-fcn.sh
@@ -28,7 +28,7 @@ pushd `dirname $0` > /dev/null
 SCRIPTPATH=`pwd`
 popd > /dev/null
 
-converter="python $SCRIPTPATH/import-caffe-dag.py"
+converter="python $SCRIPTPATH/import-caffe.py"
 data="$SCRIPTPATH/../data"
 
 mkdir -p "$data/tmp/fcn"
