+++ utils/import-ref-models.sh
@@ -18,7 +18,7 @@ pushd `dirname $0` > /dev/null
 SCRIPTPATH=`pwd`
 popd > /dev/null
 
-converter="python $SCRIPTPATH/import-caffe-dag.py"
+converter="python $SCRIPTPATH/import-caffe.py"
 data="$SCRIPTPATH/../data"
 
 mkdir -p "$data"/{tmp/vgg,tmp/caffe,models}
